refactor(precipitation-chart): improvements
