+++ src/app/app.component.ts
@@ -114,7 +114,7 @@ export class AppComponent implements OnInit {
     this.nextPrecipitationHourly = getLastXHoursForForecastInBlocks(
       point.forecast.data_1h,
       12,
-      14
+      12
     );
   }
 }
