+++ src/app/components/charts/precipitation-chart/precipitation-chart.component.ts
@@ -1,5 +1,6 @@
 import {
   Component,
+  computed,
   ElementRef,
   EventEmitter,
   Input,
@@ -14,6 +15,7 @@ import { IDataChart } from '../../../interfaces/chart.interface';
 import more from 'highcharts/highcharts-more';
 import * as Highcharts from 'highcharts';
 import { NgIf } from '@angular/common';
+import { precipitationOptionsChart } from '../../../utils/data';
 more(Highcharts);
 
 @Component({
@@ -31,14 +33,22 @@ export class PrecipitationChartComponent implements OnInit, OnDestroy {
   private chartContainer!: ElementRef;
   @Output() datesChange = new EventEmitter<Date[]>();
 
-  public chartDataArray = signal<IDataChart[][]>([]);
-  public chartData = signal<IDataChart[]>([]);
-  public currentHourBlock = signal(0);
-
   private chart!: Highcharts.Chart;
 
+  public readonly chartDataArray = computed(() =>
+    this.isDaily ? [] : (this.data as IDataChart[][])
+  );
+
+  public readonly chartData = computed(() => {
+    const currentData = this.isDaily
+      ? (this.data as IDataChart[])
+      : (this.data[this.currentHourBlock()] as IDataChart[]);
+    return currentData;
+  });
+
+  public currentHourBlock = signal(0);
+
   ngOnInit() {
-    this.chartDataArray.set(this.isDaily ? [] : (this.data as IDataChart[][]));
     this.updateChartData();
     this.createChart();
   }
@@ -75,71 +85,14 @@ export class PrecipitationChartComponent implements OnInit, OnDestroy {
   }
 
   private updateChartData() {
-    const currentData = this.isDaily
-      ? (this.data as IDataChart[])
-      : (this.data[this.currentHourBlock()] as IDataChart[]);
-    this.chartData.set(currentData);
-    this.emitDates(currentData as IDataChart[]);
+    this.emitDates(this.chartData());
   }
 
   private createChart() {
-    const options: Highcharts.Options = {
-      title: {
-        text: '',
-      },
-      plotOptions: {
-        column: {
-          pointPadding: 0,
-          borderWidth: 1,
-          groupPadding: 0,
-        },
-      },
-      xAxis: {
-        type: 'datetime',
-        ...(this.isDaily
-          ? {
-              labels: {
-                useHTML: true,
-                formatter: function () {
-                  const date = new Date(this.value);
-                  date.setDate(date.getDate() + 1);
-                  const weekday = date.toLocaleDateString('en-US', {
-                    weekday: 'short',
-                  });
-                  const formattedDate = `${String(date.getDate()).padStart(
-                    2,
-                    '0'
-                  )}/${String(date.getMonth() + 1).padStart(2, '0')}`;
-                  return `<div style="text-align: center;">
-                      <div style="font-size: 10px;">${formattedDate}</div>
-                      <div style="font-size: 14px; font-weight: bold;">${weekday}</div>
-                    </div>`;
-                },
-              },
-            }
-          : {}),
-        minTickInterval: (this.isDaily ? 24 : 1) * 3600 * 1000,
-        startOnTick: true,
-      },
-      yAxis: {
-        title: {
-          text: 'Precipitação (mm)',
-        },
-        min: 0,
-        minRange: 1,
-      },
-      colors: ['#3667a6'],
-      series: [
-        {
-          type: 'column',
-          name: 'Precipitação',
-          data: this.chartData().map((point) => [
-            point.date.getTime(),
-            point.value,
-          ]),
-        },
-      ],
-    };
+    const options: Highcharts.Options = precipitationOptionsChart(
+      this.chartData(),
+      this.isDaily
+    );
 
     this.chart = Highcharts.chart(this.chartContainer.nativeElement, options);
   }
